core(errors): class rename and re-source page load errors
