+++ lighthouse-core/gather/computed/metrics/first-cpu-idle.js
@@ -6,7 +6,7 @@
 'use strict';
 const MetricArtifact = require('./metric');
 const TracingProcessor = require('../../../lib/traces/tracing-processor');
-const LHError = require('../../../lib/errors');
+const LHError = require('../../../lib/lh-error');
 
 const LONG_TASK_THRESHOLD = 50;
