+++ lighthouse-core/gather/computed/speedline.js
@@ -7,7 +7,7 @@
 
 const ComputedArtifact = require('./computed-artifact');
 const speedline = require('speedline-core');
-const LHError = require('../../lib/errors');
+const LHError = require('../../lib/lh-error');
 
 class Speedline extends ComputedArtifact {
   get name() {
